Complete package transaction

+++ src/main/java/charging_manage_be/model/entity/service_package/PackageTransactionEntity.java
@@ -2,10 +2,9 @@ package charging_manage_be.model.entity.service_package;
 
 import charging_manage_be.model.entity.users.UserEntity;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -17,17 +16,21 @@ public class PackageTransactionEntity {
     @Id
     @Column(name = "package_transaction_id", nullable = false)
     private String packageTransactionId;
+
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
     private UserEntity user;
+
     @ManyToOne
     @JoinColumn(name = "service_package_id", nullable = false)
     private ServicePackageEntity servicePackage;
+
     @OneToOne
     @JoinColumn(name = "payment_service_package_id", nullable = false)
     private PaymentServicePackageEntity paymentServicePackage;
+
     @Column(name = "remaining_quota", nullable = false)
-    private double remainingQuota;
+    private BigDecimal remainingQuota;
     @Column(name = "status", nullable = false)
     private String status;
     @Column(name = "sign_package_at", nullable = false)
